Swap out oauth1 for oauth2
* Remove simpleoauth gem requirement
* Build bearer header for Oauth2
