+++ angellist_api.gemspec
@@ -10,7 +10,6 @@ Gem::Specification.new do |s|
   s.add_dependency 'faraday', '~> 0.7.4'
   s.add_dependency 'faraday_middleware', '~> 0.7.0'
   s.add_dependency 'multi_json', '~> 1.0.0'
-  s.add_dependency 'simple_oauth', '~> 0.1.5'
   
   s.name        = "angellist_api"
   s.version     = AngellistApi::VERSION
