+++ lib/angellist_api/authentication.rb
@@ -8,10 +8,7 @@ module AngellistApi
     # @return [Hash]
     def authentication
       {
-        :consumer_key     => respond_to?(:consumer_key) ? consumer_key : nil,
-        :consumer_secret  => respond_to?(:consumer_secret) ? consumer_secret : nil,
-        :token            => respond_to?(:oauth_token) ? oauth_token : nil,
-        :token_secret     => respond_to?(:oauth_token_secret) ? oauth_token_secret : nil,
+        :access_token => respond_to?(:access_token) ? access_token : nil,
       }
     end
