+++ lib/angellist_api/configuration.rb
@@ -6,25 +6,16 @@ module AngellistApi
     # An array of valid keys in the options hash when configuring a {AngellistApi::API}
     VALID_OPTIONS_KEYS = [
       :adapter,
-      :consumer_key,
-      :consumer_secret,
       :endpoint,
       :format,
       :gateway,
-      :oauth_token,
-      :oauth_token_secret,
+      :access_token,
       :proxy,
       :user_agent,
       :faraday_options].freeze
 
     # The adapter that will be used to connect if none is set
     DEFAULT_ADAPTER = :net_http
-
-    # By default, don't set an application key
-    DEFAULT_CONSUMER_KEY = nil
-
-    # By default, don't set an application secret
-    DEFAULT_CONSUMER_SECRET = nil
 
     # The endpoint that will be used to connect if none is set
     DEFAULT_ENDPOINT = "https://api.angel.co/".freeze
@@ -34,11 +25,8 @@ module AngellistApi
     # @note JSON is preferred over XML because it is more concise and faster to parse.
     DEFAULT_FORMAT = :json
 
-    # By default, don't set a user oauth token
-    DEFAULT_OAUTH_TOKEN = nil
-
-    # By default, don't set a user oauth secret
-    DEFAULT_OAUTH_TOKEN_SECRET = nil
+    # By default, don't set a user oauth access token
+    DEFAULT_ACCESS_TOKEN = nil
 
     # By default, don't use a proxy server
     DEFAULT_PROXY = nil
@@ -73,12 +61,9 @@ module AngellistApi
     # Reset all configuration options to defaults
     def reset
       self.adapter            = DEFAULT_ADAPTER
-      self.consumer_key       = DEFAULT_CONSUMER_KEY
-      self.consumer_secret    = DEFAULT_CONSUMER_SECRET
       self.endpoint           = DEFAULT_ENDPOINT
       self.format             = DEFAULT_FORMAT
-      self.oauth_token        = DEFAULT_OAUTH_TOKEN
-      self.oauth_token_secret = DEFAULT_OAUTH_TOKEN_SECRET
+      self.access_token       = DEFAULT_ACCESS_TOKEN
       self.proxy              = DEFAULT_PROXY
       self.user_agent         = DEFAULT_USER_AGENT
       self.gateway            = DEFAULT_GATEWAY
