+++ lib/angellist_api/connection.rb
@@ -2,7 +2,7 @@ require 'faraday_middleware'
 require 'faraday/request/phoenix'
 require 'faraday/request/multipart_with_file'
 require 'faraday/request/gateway'
-require 'faraday/request/twitter_oauth'
+require 'faraday/request/angellist_api_oauth'
 require 'faraday/response/raise_http_4xx'
 require 'faraday/response/raise_http_5xx'
