--- lib/faraday/request/twitter_oauth.rb
@@ -1,24 +1,0 @@
-require 'faraday'
-
-module Faraday
-  class Request::AngellistApiOAuth < Faraday::Middleware
-    dependency 'simple_oauth'
-
-    def call(env)
-      params = env[:body] || {}
-      signature_params = params
-
-      params.map{ |k,v| signature_params = {} if v.respond_to?(:content_type) }
-
-      header = SimpleOAuth::Header.new(env[:method], env[:url], signature_params, @options)
-
-      env[:request_headers]['Authorization'] = header.to_s
-
-      @app.call(env)
-    end
-
-    def initialize(app, options)
-      @app, @options = app, options
-    end
-  end
-end
