+++ spec/lib/angellist_api/authentication_spec.rb
@@ -7,7 +7,7 @@ describe AngellistApi::Authentication do
   
   class FullClass
     include AngellistApi::Authentication
-    attr_accessor :consumer_key, :consumer_secret, :oauth_token, :oauth_token_secret
+    attr_accessor :access_token
   end
 
   describe "#authentication" do
@@ -28,10 +28,7 @@ describe AngellistApi::Authentication do
       end
 
       it "should return a hash with nil values if auth variables are set" do
-        @a.consumer_key = "key"
-        @a.consumer_secret = "secret"
-        @a.oauth_token = "token"
-        @a.oauth_token_secret = "token_secret"
+        @a.access_token = "token"
         @a.send(:authentication).values.all?.should be_true
       end
     end
@@ -41,19 +38,14 @@ describe AngellistApi::Authentication do
     before(:each) do
       @a = FullClass.new
     end
-    
-    it "should return false if authentication has all nil values" do
-      @a.expects(:authentication).returns({:consumer_key=>nil, :consumer_secret=>nil, :oauth_token=>nil, :oauth_token_secret=>nil})
-      @a.send(:authenticated?).should be_false
-    end
-    
+        
     it "should return false if authentication has any nil values" do
-      @a.expects(:authentication).returns({:consumer_key=>nil, :consumer_secret=>"2", :oauth_token=>"3", :oauth_token_secret=>"4"})
+      @a.expects(:authentication).returns({:access_token=>nil})
       @a.send(:authenticated?).should be_false
     end
     
     it "should return true if authentication has no nil values" do
-      @a.expects(:authentication).returns({:consumer_key=>"1", :consumer_secret=>"2", :oauth_token=>"3", :oauth_token_secret=>"4"})
+      @a.expects(:authentication).returns({:access_token=>"1"})
       @a.send(:authenticated?).should be_true
     end
   end
